+++ sphinx/source/changelog.rst
@@ -14,6 +14,8 @@ Changelog
   volume)
 * Added the ``--ay-mode`` option to :ref:`trace.py` (for setting the AY stereo
   mode when writing a WAV file)
+* Added the ``aymode`` parameter to the :ref:`AUDIO` macro (to set the AY
+  stereo mode)
 * Added the ``--ay-res`` option to :ref:`trace.py` (for setting the AY sampling
   resolution when writing a WAV file)
 * Added the ``ayres`` parameter to the :ref:`AUDIO` macro (to set the AY